requirements.txt
matplotlib==3.6.1
numpy==1.23.4
scipy==1.9.3
Pillow~=9.2.0
opencv-python~=4.6.0.66
pycuda~=2022.1+cuda116
pytools~=2022.1.12
Mako~=1.2.3
setuptools~=60.2.0